Fix neighbor list dtype mismatch with float32 systems
The numpy backend always returns float64 arrays. torch.as_tensor(D)
preserved that dtype, causing system.add_neighbor_list() to reject
the neighbors when the system is float32.

Cast the distance vectors to the system's positions dtype, matching
the behavior of the torch backend which already does this correctly.

+++ python/vesin/vesin/metatomic/_neighbors.py
@@ -123,7 +123,7 @@ class NeighborList:
         )
         P = torch.as_tensor(P, dtype=torch.int32)
         S = torch.as_tensor(S, dtype=torch.int32)
-        D = torch.as_tensor(D)
+        D = torch.as_tensor(D, dtype=points.dtype)
 
         self._components = self._components.to(device=D.device)
         self._properties = self._properties.to(device=D.device)
